# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sage Oak App Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /raci/rows:
    post:
      operationId: createRaciRow
      tags: [raci]
      summary: Add a RACI row (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciRowInput"
      responses:
        "200":
          description: Created row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciRow"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /raci/rows/{id}:
    patch:
      operationId: updateRaciRow
      tags: [raci]
      summary: Rename, recategorize, or (un)link a RACI row (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciRowUpdate"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciRow"
        "404":
          description: Row not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: The row name was changed by another admin since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciNameConflict"
    delete:
      operationId: deleteRaciRow
      tags: [raci]
      summary: Remove a RACI row (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: expectedName
          in: query
          required: false
          description: >-
            The row name the client last saw. When provided and it no longer
            matches the stored name, the delete is rejected with 409 so a
            concurrent admin's rename is not silently destroyed.
          schema:
            type: string
        - name: expectedAssignments
          in: query
          required: false
          description: >-
            Canonical fingerprint of the row's cell assignments the client
            last saw: "memberId=value" pairs sorted by memberId and joined
            with commas (empty string when the row had no assignments). When
            provided and it no longer matches the stored assignments, the
            delete is rejected with 409 so a concurrent admin's assignment
            changes are not silently destroyed.
          schema:
            type: string
      responses:
        "200":
          description: Row removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Row not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: >-
            The row's name or assignments were changed by another admin since
            it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciNameConflict"
  /raci/members:
    post:
      operationId: createRaciMember
      tags: [raci]
      summary: Add a team member column (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciMemberInput"
      responses:
        "200":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciMember"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /raci/members/{id}:
    patch:
      operationId: updateRaciMember
      tags: [raci]
      summary: Rename a team member column (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciMember"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: The member name was changed by another admin since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciNameConflict"
    delete:
      operationId: deleteRaciMember
      tags: [raci]
      summary: Remove a team member column (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: expectedName
          in: query
          required: false
          description: >-
            The member name the client last saw. When provided and it no
            longer matches the stored name, the delete is rejected with 409 so
            a concurrent admin's rename is not silently destroyed.
          schema:
            type: string
        - name: expectedAssignments
          in: query
          required: false
          description: >-
            Canonical fingerprint of the member's column assignments the
            client last saw: "rowId=value" pairs sorted by rowId and joined
            with commas (empty string when the member had no assignments).
            When provided and it no longer matches the stored assignments, the
            delete is rejected with 409 so a concurrent admin's assignment
            changes are not silently destroyed.
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: >-
            The member's name or assignments were changed by another admin
            since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciNameConflict"
  /raci/cells:
    put:
      operationId: setRaciCell
      tags: [raci]
      summary: Set or clear a RACI cell (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciCellInput"
      responses:
        "200":
          description: Cell state after the update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciCellState"
        "404":
          description: Row or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: The cell was changed by another admin since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciCellConflict"
  /raci/teams/{id}/rename-category:
    post:
      operationId: renameRaciCategory
      tags: [raci]
      summary: Rename a category within a team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaciCategoryRename"
      responses:
        "200":
          description: Category renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: >-
            The category no longer exists on this team — it was likely renamed
            or removed by another admin since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaciCategoryRenameConflict"
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    RaciValue:
      type: string
    RaciAssignment:
      type: object
      required: [memberId, value]
      properties:
        memberId: { type: integer }
        value:
          $ref: "#/components/schemas/RaciValue"
    RaciMember:
      type: object
      required: [id, teamId, name, sortOrder]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        name: { type: string }
        userId:
          type: ["integer", "null"]
        sortOrder: { type: integer }
    RaciRow:
      type: object
      required: [id, teamId, name, sortOrder, assignments]
      properties:
        id: { type: integer }
        teamId: { type: integer }
        category:
          type: ["string", "null"]
        name: { type: string }
        sortOrder: { type: integer }
        applicationId:
          type: ["integer", "null"]
        appName:
          type: ["string", "null"]
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/RaciAssignment"
...
    RaciRowInput:
      type: object
      required: [teamId, name]
      properties:
        teamId: { type: integer }
        name:
          type: string
          minLength: 1
        category:
          type: ["string", "null"]
    RaciRowUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        expectedName:
          description: >-
            The row name the client last saw. When provided with a rename and
            it no longer matches the stored name, the update is rejected with
            409 so a concurrent admin's rename is not silently overwritten.
          type: string
        category:
          type: ["string", "null"]
        applicationId:
          type: ["integer", "null"]
    RaciMemberInput:
      type: object
      required: [teamId, name]
      properties:
        teamId: { type: integer }
        name:
          type: string
          minLength: 1
    RaciMemberUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        expectedName:
          description: >-
            The member name the client last saw. When provided and it no
            longer matches the stored name, the update is rejected with 409 so
            a concurrent admin's rename is not silently overwritten.
          type: string
    RaciNameConflict:
      type: object
      required: [message, currentName]
      properties:
        message: { type: string }
        currentName: { type: string }
    RaciCellInput:
      type: object
      required: [rowId, memberId, value]
      properties:
        rowId: { type: integer }
        memberId: { type: integer }
        value:
          oneOf:
            - $ref: "#/components/schemas/RaciValue"
            - type: "null"
        expectedValue:
          description: >-
            The cell value the client last saw. When provided and it no longer
            matches the stored value, the update is rejected with 409 so a
            concurrent admin's edit is not silently overwritten.
          oneOf:
            - $ref: "#/components/schemas/RaciValue"
            - type: "null"
    RaciCellConflict:
      type: object
      required: [message, currentValue]
      properties:
        message: { type: string }
        currentValue:
          oneOf:
            - $ref: "#/components/schemas/RaciValue"
            - type: "null"
    RaciCategoryRenameConflict:
      type: object
      required: [message, removed]
      properties:
        message: { type: string }
        removed:
          type: boolean
          description: True when the category no longer exists on the team
        currentName:
          type: string
          description: The category's current name, when it could be discovered
    RaciCellState:
      type: object
      required: [rowId, memberId, value]
      properties:
        rowId: { type: integer }
        memberId: { type: integer }
        value:
          oneOf:
            - $ref: "#/components/schemas/RaciValue"
            - type: "null"
    RaciCategoryRename:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          minLength: 1
        to:
          type: string
          minLength: 1
...